Moving Files to the correct app.
